+++ crates/swc_ecma_minifier/scripts/update-list.sh
@@ -11,5 +11,27 @@ set -eu
 # Clean the ignore list
 echo '' > tests/TODO.txt
 
-scripts/add-golden.sh
-scripts/ignore.sh 'fixture_tests__terser__compress__'
+unset SWC_RUN
+unset SWC_CHECK
+
+# Update golden.txt
+cargo test --test compress fixture_tests__terser__compress__ \
+  | grep 'fixture_tests__terser__compress__' \
+  | grep 'js .\.\. ok$' \
+  | sed -e 's!test fixture_tests__terser__compress__!!' \
+  | sed -e 's! ... ok!!' \
+  | sed -e 's!__!/!g' \
+  | sed -e 's!_js!.js!' \
+  >> tests/golden.txt
+
+# Update TODO.txt
+cargo test --test compress 'fixture_tests__terser__compress__' \
+  | grep 'fixture_tests__terser__compress__' \
+  | grep 'js .\.\. FAILED$' \
+  | sed -e 's!test fixture_tests__terser__compress__!!' \
+  | sed -e 's! ... FAILED!!' \
+  | sed -e 's!__!/!g' \
+  | sed -e 's!_js!.js!' \
+  >> tests/TODO.txt
+
+./scripts/sort.sh