+++ crates/swc_ecma_minifier/tests/compress.rs
@@ -51,7 +51,12 @@ fn load_txt(filename: &str) -> Vec<String> {
 }
 
 fn is_ignored(path: &Path) -> bool {
-    static IGNORED: Lazy<Vec<String>> = Lazy::new(|| load_txt("tests/TODO.txt"));
+    static IGNORED: Lazy<Vec<String>> = Lazy::new(|| {
+        load_txt("tests/TODO.txt")
+            .into_iter()
+            .chain(load_txt("tests/postponed.txt"))
+            .collect()
+    });
 
     static GOLDEN: Lazy<Vec<String>> = Lazy::new(|| load_txt("tests/golden.txt"));