+++ crates/swc_ecma_parser/src/parser/class_and_fn.rs
@@ -195,8 +195,12 @@ impl<'a, I: Tokens> Parser<I> {
     where
         T: OutputType,
     {
-        let (ident, mut class) =
-            self.parse_class_inner(start, class_start, decorators, T::IS_IDENT_REQUIRED)?;
+        let (ident, mut class) = self
+            .with_ctx(Context {
+                in_class: true,
+                ..self.ctx()
+            })
+            .parse_class_inner(start, class_start, decorators, T::IS_IDENT_REQUIRED)?;
 
         if is_abstract {
             class.is_abstract = true
@@ -1278,6 +1282,11 @@ impl<'a, I: Tokens> Parser<I> {
             in_async: is_async,
             in_generator: is_generator,
             in_arrow_function: is_arrow_function,
+            inside_non_arrow_function_scope: if is_arrow_function {
+                self.ctx().inside_non_arrow_function_scope
+            } else {
+                true
+            },
             in_function: true,
             is_break_allowed: false,
             is_continue_allowed: false,