+++ crates/swc_ecma_parser/tests/typescript-errors/meta-property/new-target/in-arrow-function/input.ts
@@ -1,0 +1,3 @@
+const f = () => {
+    new.target;
+};