+++ crates/swc_ecma_parser/tests/typescript/meta-property/new-target/in-class/input.ts
@@ -1,0 +1,13 @@
+class C1 {
+    constructor() {
+        new.target;
+    }
+}
+
+class C2 {
+    constructor() {
+        const a = () => {
+            new.target;
+        };
+    }
+}